FJLIB Using External for FJ/Pnodes
Fjlib where fjnodeX and pnodeX is external and resolved by the task program.
X Being the number of core.

Main.c in the taskset will define the external variables that the linker will resolve when compiling.

Unsure if this is a viable solution as results differ from without using the library.